--- packages/middleware-form-data/CHANGELOG.md
@@ -1,15 +1,0 @@
-## v1.4.0
-
--   🐛 No longer transform `payload` when it is `FormData` ([#153](https://github.com/fanhaoyuan/fatcher/pull/153))
-
-## v1.2.0
-
--   🔧 Using `requestHeaders` instead of `headers` in context ([#142](https://github.com/fanhaoyuan/fatcher/pull/142))
-
-## v1.0.1
-
--   🐛 Should not return request headers but response headers.
-
-## v1.0.0
-
--   🚀 Supports `FormData` body.
